service_master/requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
Deprecated==1.2.14
exceptiongroup==1.2.0
fastapi==0.105.0
frozenlist==1.4.0
h11==0.14.0
httptools==0.6.1
idna==3.6
importlib-resources==6.1.1
limits==3.7.0
multidict==6.0.4
packaging==23.2
pydantic==2.5.2
pydantic_core==2.14.5
pylogbeat==2.0.1
python-dotenv==1.0.0
python-logstash-async==2.7.2
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.9.0
urllib3==2.1.0
uvicorn==0.24.0.post1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
wrapt==1.16.0
yarl==1.9.4
